refactor(podman): extract VirtualMachinePlatformCheck class from `podman-install.ts`
