+++ extensions/podman/packages/extension/src/utils/podman-install.spec.ts
@@ -30,7 +30,6 @@ import {
   getBundledPodmanVersion,
   HyperVCheck,
   PodmanInstall,
-  VirtualMachinePlatformCheck,
   WinInstaller,
   WSL2Check,
   WSLVersionCheck,
@@ -317,57 +316,6 @@ test('expect winMemory preflight check return failure result if the machine has 
   expect(result.docLinksDescription).toBeUndefined();
   expect(result.docLinks).toBeUndefined();
   expect(result.fixCommand).toBeUndefined();
-});
-
-test('expect winVirtualMachine preflight check return successful result if the virtual machine platform feature is enabled', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() =>
-    Promise.resolve({
-      stdout: 'VirtualMachinePlatform',
-      stderr: '',
-      command: 'command',
-    }),
-  );
-
-  const winVirtualMachinePlatformCheck = new VirtualMachinePlatformCheck();
-  const result = await winVirtualMachinePlatformCheck.execute();
-  expect(extensionApi.process.exec).toBeCalledWith(expect.anything(), expect.arrayContaining([expect.anything()]), {
-    encoding: 'utf16le',
-  });
-  expect(result.successful).toBeTruthy();
-});
-
-test('expect winVirtualMachine preflight check return successful result if the virtual machine platform feature is disabled', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() =>
-    Promise.resolve({
-      stdout: 'some message',
-      stderr: '',
-      command: 'command',
-    }),
-  );
-
-  const winVirtualMachinePlatformCheck = new VirtualMachinePlatformCheck();
-  const result = await winVirtualMachinePlatformCheck.execute();
-  expect(result.description).equal('Virtual Machine Platform should be enabled to be able to run Podman.');
-  expect(result.docLinksDescription).equal('Learn about how to enable the Virtual Machine Platform feature:');
-  expect(result.docLinks?.[0].url).equal(
-    'https://learn.microsoft.com/en-us/windows/wsl/install-manual#step-3---enable-virtual-machine-feature',
-  );
-  expect(result.docLinks?.[0].title).equal('Enable Virtual Machine Platform');
-});
-
-test('expect winVirtualMachine preflight check return successful result if there is an error when checking if virtual machine platform feature is enabled', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(() => {
-    throw new Error();
-  });
-
-  const winVirtualMachinePlatformCheck = new VirtualMachinePlatformCheck();
-  const result = await winVirtualMachinePlatformCheck.execute();
-  expect(result.description).equal('Virtual Machine Platform should be enabled to be able to run Podman.');
-  expect(result.docLinksDescription).equal('Learn about how to enable the Virtual Machine Platform feature:');
-  expect(result.docLinks?.[0].url).equal(
-    'https://learn.microsoft.com/en-us/windows/wsl/install-manual#step-3---enable-virtual-machine-feature',
-  );
-  expect(result.docLinks?.[0].title).equal('Enable Virtual Machine Platform');
 });
 
 test('expect WSLVersion preflight check return fail result if wsl --version command fails its execution', async () => {
